La til session, og logg inn for kunder

+++ frontend/app.py
@@ -1,11 +1,13 @@
-from flask import Flask, render_template, request, redirect, url_for
+from flask import Flask, render_template, request, redirect, url_for, session
 from flask_cors import CORS
 import requests
+import secrets
 
 # Setter opp applikasjon
 app = Flask(__name__)
 CORS(app)
 
+app.secret_key = secrets.token_hex()
 
 # Rute for hjemmeside
 @app.route('/', methods = ["GET"])
@@ -36,10 +38,15 @@ def logg_inn():
     passord = request.form.get('passord')
     bruker = requests.get('http://127.0.0.1:5010/logg_inn', json={"navn": navn, "passord": passord}).json()
 
-    if bruker["status"] == "finnes":
-      return redirect(url_for('eier_sin_side', rid=bruker["rid"], navn=bruker["navn"]))
-    
-    print(bruker["status"])
+    # Sjekker om bruker er eier
+    if bruker["status"] == "finnes" and bruker["rid"] != None:
+      session["user"] = navn
+      session["rid"] = bruker["rid"]
+      return redirect(url_for('eier_sin_side', rid=session["rid"], navn=session["user"]))
+    # Sjekker om bruker er kunde
+    if bruker["status"] == "finnes" and bruker["rid"] == None:
+      session["user"] = navn
+      return redirect(url_for('kunde_sin_side.html', navn=session["user"]))
     return render_template('logg_inn.html', status=bruker["status"])
 
 
@@ -47,58 +54,88 @@ def logg_inn():
 # rute som sender bruker til eier_sin_side.html
 @app.route('/eier_sin_side/<rid>/<navn>', methods = ["GET"])
 def eier_sin_side(rid, navn):
-  restaurant_meny = requests.get('http://127.0.0.1:5010/get_restaurant_meny', json={"rid": rid}).json()
-  restaurant = requests.get('http://127.0.0.1:5010/get_restaurant', json={"rid": rid}).json()
-  return render_template('eier_sin_side.html', navn=navn, meny=restaurant_meny, restaurant=restaurant, rid=restaurant[0])
+  if "user" in session and "rid" in session:
+    restaurant_meny = requests.get('http://127.0.0.1:5010/get_restaurant_meny', json={"rid": rid}).json()
+    restaurant = requests.get('http://127.0.0.1:5010/get_restaurant', json={"rid": rid}).json()
+    return render_template('eier_sin_side.html', navn=navn, meny=restaurant_meny, restaurant=restaurant, rid=restaurant[0])
+  else:
+    return redirect('/logg_inn')
 
 
 
 # Rute som sender bruker til legg_til.html side
 @app.route('/legg_til_ny_rett/<rid>/<navn>', methods=["GET"])
 def legg_til_ny_rett(rid, navn):
-  return render_template('legg_til.html', rid=rid, navn=navn)
+  if "user" in session and "rid" in session:
+    return render_template('legg_til.html', rid=rid, navn=navn)
+  else:
+    return redirect('/logg_inn')
+
 
 
 # Rute som sender request til serveren for å fjerne valgt rett
 @app.route('/fjern_rett/<rett_id>/<rid>/<navn>', methods=["POST"])
 def fjern_rett(rett_id, rid, navn):
-  requests.delete('http://127.0.0.1:5010/fjern_rett', json={"rett_id": rett_id, "rid": rid})
-  return redirect(url_for('eier_sin_side', rid=rid, navn=navn))
+  if "user" in session and "rid" in session:
+    requests.delete('http://127.0.0.1:5010/fjern_rett', json={"rett_id": rett_id, "rid": rid})
+    return redirect(url_for('eier_sin_side', rid=rid, navn=navn))
+  else:
+    return redirect('/logg_inn')
 
 
 
 # Rute som sender bruker til rediger.html side
 @app.route('/rediger_rett/<rett_id>/<rid>/<navn>', methods=["get"])
 def rediger_rett(rett_id, rid, navn):
-  rett = requests.get('http://127.0.0.1:5010/get_rett', json={"rett_id": rett_id, "rid": rid}).json()
-  return render_template('rediger.html', rett=rett, rett_id=rett_id, rid=rid, navn=navn)
+  if "user" in session and "rid" in session:
+    rett = requests.get('http://127.0.0.1:5010/get_rett', json={"rett_id": rett_id, "rid": rid}).json()
+    return render_template('rediger.html', rett=rett, rett_id=rett_id, rid=rid, navn=navn)
+  else:
+    return redirect('/logg_inn')
 
 
 
 # Rute for å endre navn på retten
 @app.route('/rediger/rett/<rett_id>/<rid>/<navn>', methods=["POST"])
 def rediger_rett_tekst(rett_id, rid, navn):
-  rett_tekst = request.form.get('tekst')
-  requests.put('http://127.0.0.1:5010/rediger_tekst', json={"rett_id": rett_id, "rid": rid, "content": rett_tekst})
-  return redirect(url_for('eier_sin_side', rid=rid, navn=navn))
+  if "user" in session and "rid" in session:
+    rett_tekst = request.form.get('tekst')
+    requests.put('http://127.0.0.1:5010/rediger_tekst', json={"rett_id": rett_id, "rid": rid, "content": rett_tekst})
+    return redirect(url_for('eier_sin_side', rid=rid, navn=navn))
+  else:
+    return redirect('/logg_inn')
 
 
 
 # Rute for å endre beskrivelse på retten
 @app.route('/rediger/beskrivelse/<rett_id>/<rid>/<navn>', methods=["POST"])
 def rediger_rett_beskrivelse(rett_id, rid, navn):
-  rett_beskrivelse = request.form.get('beskrivelse')
-  requests.put('http://127.0.0.1:5010/rediger_beskrivelse', json={"rett_id": rett_id, "rid": rid, "content": rett_beskrivelse})
-  return redirect(url_for('eier_sin_side', rid=rid, navn=navn))
+  if "user" in session and "rid" in session:
+    rett_beskrivelse = request.form.get('beskrivelse')
+    requests.put('http://127.0.0.1:5010/rediger_beskrivelse', json={"rett_id": rett_id, "rid": rid, "content": rett_beskrivelse})
+    return redirect(url_for('eier_sin_side', rid=rid, navn=navn))
+  else:
+    return redirect('/logg_inn')
 
 
 
 # Rute for å endre pris på retten
 @app.route('/rediger/pris/<rett_id>/<rid>/<navn>', methods=["POST"])
 def rediger_rett_pris(rett_id, rid, navn):
-  rett_pris = request.form.get('pris')
-  requests.put('http://127.0.0.1:5010/rediger_pris', json={"rett_id": rett_id, "rid": rid, "content": rett_pris})
-  return redirect(url_for('eier_sin_side', rid=rid, navn=navn))
+  if "user" in session and "rid" in session:
+    rett_pris = request.form.get('pris')
+    requests.put('http://127.0.0.1:5010/rediger_pris', json={"rett_id": rett_id, "rid": rid, "content": rett_pris})
+    return redirect(url_for('eier_sin_side', rid=rid, navn=navn))
+  else:
+    return redirect('/logg_inn')
+
+
+
+# Rute for å logge ut
+@app.route('/logg_ut')
+def logg_ut():
+    session.pop('user', None)
+    return redirect(url_for('index'))
